Internal.CloseHandle: ReleaseHandle() should not check close(2) return value
According to the manual page for close(2), on Linux, irregardless of
the error returned by the system call, the file descriptor will always
be closed.  Checking if the syscall failed shouldn't affect program
behavior unless it's for diagnostic or remedial purposes (and there
isn't much to be done when we're talking about socket file descriptors
here).